+++ PostHogTests/PostHogFeatureFlagsTest.swift
@@ -11,11 +11,12 @@ import Nimble
 import Quick
 
 class PostHogFeatureFlagsTest: QuickSpec {
-    func getSut() -> PostHogFeatureFlags {
-        let config = PostHogConfig(apiKey: "123", host: "http://localhost:9001")
-        let storage = PostHogStorage(config)
+    let config = PostHogConfig(apiKey: "123", host: "http://localhost:9001")
+
+    func getSut(storage: PostHogStorage? = nil) -> PostHogFeatureFlags {
+        let theStorage = storage ?? PostHogStorage(config)
         let api = PostHogApi(config)
-        return PostHogFeatureFlags(config, storage, api)
+        return PostHogFeatureFlags(config, theStorage, api)
     }
 
     override func spec() {
@@ -136,5 +137,76 @@ class PostHogFeatureFlagsTest: QuickSpec {
             expect(sut.isFeatureEnabled("new-flag")) == true
             expect(sut.isFeatureEnabled("bool-value")) == true
         }
+
+        #if os(iOS)
+            it("returns isSessionReplayFlagActive true if there is a value") {
+                let storage = PostHogStorage(self.config)
+
+                let recording: [String: Any] = ["test": 1]
+                storage.setDictionary(forKey: .sessionReplay, contents: recording)
+
+                let sut = self.getSut(storage: storage)
+
+                expect(sut.isSessionReplayFlagActive()) == true
+
+                storage.reset()
+            }
+
+            it("returns isSessionReplayFlagActive false if there is no value") {
+                let sut = self.getSut()
+
+                expect(sut.isSessionReplayFlagActive()) == false
+            }
+
+            it("returns isSessionReplayFlagActive false if feature flag disabled") {
+                let storage = PostHogStorage(self.config)
+
+                let recording: [String: Any] = ["test": 1]
+                storage.setDictionary(forKey: .sessionReplay, contents: recording)
+
+                let sut = self.getSut(storage: storage)
+
+                expect(sut.isSessionReplayFlagActive()) == true
+
+                let group = DispatchGroup()
+                group.enter()
+
+                sut.loadFeatureFlags(distinctId: "distinctId", anonymousId: "anonymousId", groups: ["group": "value"], callback: {
+                    group.leave()
+                })
+
+                group.wait()
+
+                expect(storage.getDictionary(forKey: .sessionReplay) == nil)
+                expect(sut.isSessionReplayFlagActive()) == false
+
+                storage.reset()
+            }
+
+            it("returns isSessionReplayFlagActive true if feature flag active") {
+                let storage = PostHogStorage(self.config)
+
+                let sut = self.getSut(storage: storage)
+
+                expect(sut.isSessionReplayFlagActive()) == false
+
+                let group = DispatchGroup()
+                group.enter()
+
+                server.returnReplay = true
+
+                sut.loadFeatureFlags(distinctId: "distinctId", anonymousId: "anonymousId", groups: ["group": "value"], callback: {
+                    group.leave()
+                })
+
+                group.wait()
+
+                expect(storage.getDictionary(forKey: .sessionReplay)) != nil
+                expect(self.config.snapshotEndpoint) == "/newS/"
+                expect(sut.isSessionReplayFlagActive()) == true
+
+                storage.reset()
+            }
+        #endif
     }
 }